EC: Basic CRUD functionality finally works

+++ src/main/java/com/qa/ims/persistence/domain/Item.java
@@ -47,7 +47,7 @@ public class Item {
 	}
 
 
-	public double getItemPrice() {
+	public Long getItemPrice() {
 		return itemPrice;
 	}
